Refactor Controls component out of application.js
