+++ app/application.js
@@ -1,6 +1,5 @@
 import Cycle from '@cycle/core';
-import { makeDOMDriver, div, h1, small, button } from '@cycle/dom';
-import isolate from '@cycle/isolate';
+import { makeDOMDriver, div, button, h1, small } from '@cycle/dom';
 import storageDriver from '@cycle/storage';
 import { Observable } from 'rx';
 import R from 'ramda';
@@ -9,20 +8,10 @@ import { trelloSinkDriver } from './drivers/Trello';
 import { makeGraphDriver } from './drivers/Graph';
 import { exportToCSVDriver } from './drivers/ExportToCSV';
 
-import LabeledSelect from './components/LabeledSelect/LabeledSelect';
-import LabeledCheckbox from './components/LabeledCheckbox/LabeledCheckbox';
-import LabeledDatePicker from './components/LabeledDatePicker/LabeledDatePicker';
-import SelectDatesButton from './components/SelectDatesButton/SelectDatesButton';
+import Controls from './components/Controls/Controls';
 import TrelloCFD from './components/TrelloCFD/TrelloCFD';
 import TrelloKanbanMetrics from './components/TrelloKanbanMetrics/TrelloKanbanMetrics';
 
-import {
-  lastMonth,
-  endOfLastMonth,
-  currentMonth,
-  today,
-  tomorrow,
-} from './utils/date';
 import { getDisplayedLists } from './utils/trello';
 import { argsToArray } from './utils/function';
 
@@ -34,172 +23,39 @@ function main({ DOMControls, DOMMetrics, TrelloFetch, TrelloMissingInfo, Storage
   const trelloLists$ = publishedTrelloLists$.startWith([]);
   const trelloActions$ = publishedTrelloActions$.startWith([]);
 
-  // Checkbox to preview tomorrow CFD
+  // Controls that configure the analysis.
 
-  const previewTomorrow = isolate(LabeledCheckbox)({
+  const controls = Controls({
     DOM: DOMControls,
-    props$: Observable.of({
-      name: 'preview-tomorrow',
-      label: 'Preview tomorrow CFD (include today operations)',
-    }),
+    boards$: TrelloFetch.boards$,
+    lists$: trelloLists$,
+    Storage,
   });
 
-  // Selects to choose displayed board and lists
-
-  const boardSelectProps$ = Observable.combineLatest(
-    TrelloFetch.boards$,
-    Storage.local.getItem('selectedBoard').first(),
-    (boards, storedBoard) => ({
-      name: 'board',
-      label: 'Board',
-      classNames: ['browser-default'],
-      select: (values) => R.defaultTo(R.head(values), storedBoard),
-
-      // Render board names instead of shortLink when possible.
-      render: (shortLink) => R.propOr(
-        shortLink,
-        'name',
-        R.find(R.propEq('shortLink', shortLink), boards)
-      ),
-    })
-  );
-
-  const boardSelect = isolate(LabeledSelect)({
-    DOM: DOMControls,
-    props$: boardSelectProps$,
-    values$: TrelloFetch.boards$.map(R.pluck('shortLink')),
-  });
-
-  const firstDisplayedListSelect = isolate(LabeledSelect)({
-    DOM: DOMControls,
-    props$: Storage.local.getItem('firstDisplayedList').first()
-      .map((storedList) => ({
-        name: 'first-displayed-list',
-        label: 'Work begins',
-        classNames: ['browser-default'],
-        select: R.cond([
-          [R.contains(storedList), R.always(storedList)],
-          [R.T, R.head],
-        ]),
-      })),
-    values$: trelloLists$.map(R.pluck('name')),
-  });
-
-  const lastDisplayedListSelect = isolate(LabeledSelect)({
-    DOM: DOMControls,
-    props$: Storage.local.getItem('lastDisplayedList').first()
-      .map((storedList) => ({
-        name: 'last-displayed-list',
-        label: 'Work ends',
-        classNames: ['browser-default'],
-        select: R.cond([
-          [R.contains(storedList), R.always(storedList)],
-          [R.T, R.last],
-        ]),
-      })),
-    values$: trelloLists$.map(R.pluck('name')),
-  });
-
-  // Buttons to select period
-
-  const selectLastMonthButton = isolate(SelectDatesButton)({
-    DOM: DOMControls,
-    props$: Observable.of({
-      label: 'Last month',
-      classNames: ['btn waves-effect waves-light trello-blue'],
-      startDate: lastMonth,
-      endDate: endOfLastMonth,
-    }),
-  });
-
-  const selectCurrentMonthButton = isolate(SelectDatesButton)({
-    DOM: DOMControls,
-    props$: Observable.of({
-      label: 'Current month',
-      classNames: ['btn waves-effect waves-light trello-blue'],
-      startDate: currentMonth,
-      endDate: today,
-    }),
-  });
-
-  const selectPeriod$ = Observable.concat(
-    Observable.combineLatest(
-      Storage.local.getItem('selectedStartDate').first()
-        .map(R.defaultTo(currentMonth)),
-      Storage.local.getItem('selectedEndDate').first()
-        .map(R.defaultTo(today)),
-      (startDate, endDate) => ({ startDate, endDate })
-    ),
-    Observable.merge(
-      selectLastMonthButton.dates$,
-      selectCurrentMonthButton.dates$
-    )
-  );
-
-  // Datepickers to select dates
-
-  const startDatePicker = isolate(LabeledDatePicker)({
-    DOM: DOMControls,
-    props$: Observable.of({
-      name: 'start-date',
-      label: 'Start Date',
-      max: today,
-    }),
-    value$: selectPeriod$.map(R.prop('startDate')),
-  });
-
-  const endDatePicker = isolate(LabeledDatePicker)({
-    DOM: DOMControls,
-    props$: Observable.of({
-      name: 'end-date',
-      label: 'End Date',
-      max: today,
-    }),
-    value$: selectPeriod$.map(R.prop('endDate')),
-  });
-
-  // Trello CFD
-
-  const parseSelectedDate = R.cond([
-    [R.isEmpty, R.always(null)],
-    [R.T, R.identity],
-  ]);
-
-  const selectedDates$ = Observable.combineLatest(
-    startDatePicker.selected$.map(parseSelectedDate),
-    endDatePicker.selected$.map(parseSelectedDate),
-    previewTomorrow.checked$,
-    (startSelected, endSelected, previewTomorrowChecked) => ({
-      startDate: startSelected,
-      endDate: previewTomorrowChecked ? tomorrow : endSelected,
-    })
-  );
+  // Compute displayed lists from controls.
 
   const trelloDisplayedLists$ = Observable.combineLatest(
     trelloLists$,
-    firstDisplayedListSelect.selected$,
-    lastDisplayedListSelect.selected$,
+    controls.selectedFirstList$,
+    controls.selectedLastList$,
     getDisplayedLists
   );
 
+  // Trello CFD
+
   const trelloCFD = TrelloCFD({
-    DOM: DOMControls,
     actions$: trelloActions$,
     lists$: trelloLists$,
     displayedLists$: trelloDisplayedLists$,
-    dates$: selectedDates$,
-    props$: Observable.of({
-      label: 'Refresh data',
-      classNames: ['btn waves-effect waves-light trello-green'],
-    }),
-    previewTomorrow$: previewTomorrow.checked$,
+    dates$: controls.selectedDates$,
+    previewTomorrow$: controls.previewTomorrow$,
   });
 
   // Trello Kanban metrics
 
   const trelloKanbanMetrics = TrelloKanbanMetrics({
     actions$: trelloActions$,
-    dates$: selectedDates$,
+    dates$: controls.selectedDates$,
     lists$: trelloDisplayedLists$,
     complementaryActions$: publishedTrelloCardsActions$$
       .switch()
@@ -217,51 +73,15 @@ function main({ DOMControls, DOMMetrics, TrelloFetch, TrelloMissingInfo, Storage
   publishedTrelloCardsActions$$.connect();
 
   return {
-    DOMControls: Observable.combineLatest(
-      boardSelect.DOM,
-      trelloCFD.DOM,
-      selectLastMonthButton.DOM,
-      selectCurrentMonthButton.DOM,
-      startDatePicker.DOM,
-      endDatePicker.DOM,
-      previewTomorrow.DOM,
-      firstDisplayedListSelect.DOM,
-      lastDisplayedListSelect.DOM,
-      (
-        boardVTree,
-        trelloCFDVTree,
-        selectLastMonthButtonVTree,
-        selectCurrentMonthButtonVTree,
-        startDatePickerVTree,
-        endDatePickerVTree,
-        previewTomorrowVTree,
-        firstDisplayedListVTree,
-        lastDisplayedListVTree
-      ) => div([
-        h1('.title.center-align.trello-blue.white-text', [
-          'TKAT ',
-          small('.trello-blue-100-text', '(Trello Kanban Analysis Tool)'),
-        ]),
-        div('.container', [
-          div('.m-top.row', [
-            div('.col.s6', [boardVTree]),
-            div('.col.s3', [firstDisplayedListVTree]),
-            div('.col.s3', [lastDisplayedListVTree]),
+    DOMControls: controls.DOM.map(
+      (controlsVTree) =>
+        div([
+          h1('.title.center-align.trello-blue.white-text', [
+            'TKAT ',
+            small('.trello-blue-100-text', '(Trello Kanban Analysis Tool)'),
           ]),
-          div('.m-top.row', [
-            div('.col.s6', [startDatePickerVTree]),
-            div('.col.s6', [endDatePickerVTree]),
-          ]),
-          div('.center-align-around', [
-            trelloCFDVTree,
-            selectLastMonthButtonVTree,
-            selectCurrentMonthButtonVTree,
-          ]),
-          div('.m-top.row', [
-            div('.col.s12', [previewTomorrowVTree]),
-          ]),
-        ]),
-      ])
+          controlsVTree,
+        ])
     ),
     DOMMetrics: trelloKanbanMetrics.DOM.map(
       (trelloKanbanMetricsVTree) =>
@@ -276,8 +96,8 @@ function main({ DOMControls, DOMMetrics, TrelloFetch, TrelloMissingInfo, Storage
         ])
     ),
     TrelloFetch: Observable.combineLatest(
-      boardSelect.selected$,
-      trelloCFD.Trello,
+      controls.selectedBoard$,
+      controls.refreshClicks$,
       R.compose(R.head, argsToArray)
     ),
     TrelloMissingInfo: trelloKanbanMetrics.Trello,
@@ -286,18 +106,7 @@ function main({ DOMControls, DOMMetrics, TrelloFetch, TrelloMissingInfo, Storage
       trelloCFD.CSV,
       R.compose(R.last, argsToArray)
     ),
-    Storage: Observable.merge(
-      boardSelect.selected$
-        .map(selected => ({ key: 'selectedBoard', value: selected })),
-      firstDisplayedListSelect.selected$
-        .map(selected => ({ key: 'firstDisplayedList', value: selected })),
-      lastDisplayedListSelect.selected$
-        .map(selected => ({ key: 'lastDisplayedList', value: selected })),
-      selectedDates$
-        .map(period => ({ key: 'selectedStartDate', value: period.startDate })),
-      selectedDates$
-        .map(period => ({ key: 'selectedEndDate', value: period.endDate }))
-    ),
+    Storage: controls.Storage,
   };
 }
